add：LeapYear文件 modify:

+++ Test/src/test002/MonthlyPayment.java
@@ -11,7 +11,7 @@ public class MonthlyPayment {
 		System.out.println("年利率（%）：");
 		double r = scanner.nextDouble() / 1200;
 		System.out.println("还款年数：");
-		double m = scanner.nextInt() * 12;
+		int m = scanner.nextInt() * 12;
 		double payment = (p * r * Math.pow((1 + r), m)) / (Math.pow((1 + r), m) - 1);
 		payment = Math.round(payment * 100) / 100.0;
 		System.out.println("每月还："+payment);
